[MaxText/tests] `pytest` now works on CPU with no special configuration ; [MaxText/globals.py] New funcs and vars for CPU-only testing ; [MaxText/max_utils.py] New funcs for guarding GCS accessibility for running tests without auth keys

+++ MaxText/tests/tokenizer_test.py
@@ -14,19 +14,22 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-""" Tests for tokenizer
-"""
+""" Tests for tokenizer """
+
+import os
+import subprocess
+import unittest
 
 import numpy as np
-from MaxText import train_tokenizer
-from MaxText.input_pipeline import _input_pipeline_utils
-from MaxText.globals import PKG_DIR
 
-import unittest
 import pytest
+
 import tensorflow_datasets as tfds
-import subprocess
-import os
+
+from MaxText import train_tokenizer
+from MaxText.globals import PKG_DIR, tpu_present
+from MaxText.input_pipeline import _input_pipeline_utils
+from MaxText.max_utils import gcs_bucket_accessible
 
 
 class TokenizerTest(unittest.TestCase):
@@ -36,6 +39,8 @@ class TokenizerTest(unittest.TestCase):
   def setUpClass(cls):
     dataset_name = "c4/en:3.0.1"
     dataset_path = "gs://maxtext-dataset"
+    if not gcs_bucket_accessible("maxtext-dataset"):
+      return
     cls.vocab_size = 32_768
     cls.max_corpus_chars = 10_000_000
     assets_path = "tests"
@@ -62,15 +67,21 @@ class TokenizerTest(unittest.TestCase):
 
   @classmethod
   def tearDownClass(cls):
-    os.remove(cls.tokenizer_path)
+    if hasattr(cls, "tokenizer_path") and os.path.isdir(cls.tokenizer_path):
+      os.remove(cls.tokenizer_path)
 
   @pytest.mark.skip(reason="mohitkhatwani@ will fix this")
   @pytest.mark.tpu_only
+  @unittest.skip("mohitkhatwani@ will fix this")
   def test_tokenize(self):
     text = "This is a test"
     self.assertTrue(np.array_equal(self.source_tokenizer.encode(text).numpy(), self.test_tokenizer.encode(text).numpy()))
 
   @pytest.mark.tpu_only
+  @unittest.skipIf(
+      not tpu_present or not gcs_bucket_accessible("maxtext-dataset"),
+      "TPU only test" if not tpu_present else "gs://maxtext-dataset bucket not accessible",
+  )
   def test_detokenize(self):
     tokens = [66, 12, 10, 698]
     self.assertEqual(np.asarray(self.source_tokenizer.decode(tokens)), np.asarray(self.test_tokenizer.decode(tokens)))
@@ -83,6 +94,8 @@ class TikTokenTest(unittest.TestCase):
   def setUpClass(cls):
     dataset_name = "c4/en:3.0.1"
     dataset_path = "gs://maxtext-dataset"
+    if not gcs_bucket_accessible("maxtext-dataset"):
+      return
     cls.source_tokenizer = _input_pipeline_utils.get_tokenizer(
         os.path.join(os.path.dirname(PKG_DIR), "assets", "tokenizer_llama3.tiktoken"),
         "tiktoken",
@@ -97,12 +110,20 @@ class TikTokenTest(unittest.TestCase):
     cls.dataset = train_ds_builder.as_dataset(split="train", read_config=read_config, shuffle_files=True)
 
   @pytest.mark.tpu_only
+  @unittest.skipIf(
+      not tpu_present or not gcs_bucket_accessible("maxtext-dataset"),
+      "TPU only test" if not tpu_present else "gs://maxtext-dataset bucket not accessible",
+  )
   def test_tokenize(self):
     text = "This is a test"
     tokens = [2028, 374, 264, 1296]
     self.assertTrue(np.array_equal(self.source_tokenizer.encode(text), tokens))
 
   @pytest.mark.tpu_only
+  @unittest.skipIf(
+      not tpu_present or not gcs_bucket_accessible("maxtext-dataset"),
+      "TPU only test" if not tpu_present else "gs://maxtext-dataset bucket not accessible",
+  )
   def test_detokenize(self):
     tokens = [2028, 374, 264, 1296]
     text = "This is a test"
@@ -114,6 +135,8 @@ class HFTokenizerTest(unittest.TestCase):
 
   @classmethod
   def setUpClass(cls):
+    if not gcs_bucket_accessible("maxtext-gemma"):
+      return
     source = "gs://maxtext-gemma/huggingface/gemma2-2b"
     destination = os.path.join(os.path.dirname(PKG_DIR), "assets")
     subprocess.run(
@@ -128,6 +151,10 @@ class HFTokenizerTest(unittest.TestCase):
     )
 
   @pytest.mark.tpu_only
+  @unittest.skipIf(
+      not tpu_present or not gcs_bucket_accessible("maxtext-gemma"),
+      "TPU only test" if not tpu_present else "gs://maxtext-gemma bucket not accessible",
+  )
   def test_tokenize(self):
     text = "This is a test"
     self.assertTrue(np.array_equal(self.hf_tokenizer.encode(text), self.sp_tokenizer.encode(text)))
